Moved class FsMultiLevelDataset into xcube.core.mldataset and added classmethod write_dataset()
