+++ test/util/test_fspath.py
@@ -4,50 +4,11 @@ import unittest
 import fsspec
 import fsspec.implementations.local
 
-from xcube.core.store.fs.helpers import get_fs_path_class
-from xcube.core.store.fs.helpers import resolve_path
+from xcube.util.fspath import get_fs_path_class
+from xcube.util.fspath import resolve_path
 
 
-class FsHelpersTest(unittest.TestCase):
-
-    def test_resolve_path_succeeds(self):
-        path = pathlib.PurePosixPath("a", "b", "c")
-        self.assertEqual(pathlib.PurePosixPath("a", "b", "c"),
-                         resolve_path(path))
-
-        path = pathlib.PurePosixPath(".", "b", "c")
-        self.assertEqual(pathlib.PurePosixPath("b", "c"),
-                         resolve_path(path))
-
-        path = pathlib.PurePosixPath("a", ".", "c")
-        self.assertEqual(pathlib.PurePosixPath("a", "c"),
-                         resolve_path(path))
-
-        path = pathlib.PurePosixPath("a", "b", ".")
-        self.assertEqual(pathlib.PurePosixPath("a", "b"),
-                         resolve_path(path))
-
-        path = pathlib.PurePosixPath("a", "..", "c")
-        self.assertEqual(pathlib.PurePosixPath("c"),
-                         resolve_path(path))
-
-        path = pathlib.PurePosixPath("a", "b", "..")
-        self.assertEqual(pathlib.PurePosixPath("a"),
-                         resolve_path(path))
-
-    def test_resolve_path_fails(self):
-        path = pathlib.PurePosixPath("..", "b", "c")
-        with self.assertRaises(ValueError) as cm:
-            resolve_path(path)
-        self.assertEqual('cannot resolve path, misplaced ".."',
-                         f'{cm.exception}')
-
-        path = pathlib.PurePosixPath("a", "..", "..")
-        with self.assertRaises(ValueError) as cm:
-            resolve_path(path)
-        self.assertEqual('cannot resolve path, misplaced ".."',
-                         f'{cm.exception}')
-
+class FsPathTest(unittest.TestCase):
     def test_get_fs_path_class(self):
         self.assertIs(pathlib.PurePath,
                       get_fs_path_class(
@@ -59,3 +20,42 @@ class FsHelpersTest(unittest.TestCase):
                       get_fs_path_class(fsspec.filesystem("s3")))
         self.assertIs(pathlib.PurePosixPath,
                       get_fs_path_class(fsspec.filesystem("memory")))
+
+    def test_resolve_path_succeeds(self):
+        path = pathlib.PurePosixPath("a", "b", "c")
+        self.assertEqual(pathlib.PurePosixPath("a", "b", "c"),
+                         resolve_path(path))
+
+        path = pathlib.PurePosixPath("../core/store/fs", "b", "c")
+        self.assertEqual(pathlib.PurePosixPath("b", "c"),
+                         resolve_path(path))
+
+        path = pathlib.PurePosixPath("a", "../core/store/fs", "c")
+        self.assertEqual(pathlib.PurePosixPath("a", "c"),
+                         resolve_path(path))
+
+        path = pathlib.PurePosixPath("a", "b", "../core/store/fs")
+        self.assertEqual(pathlib.PurePosixPath("a", "b"),
+                         resolve_path(path))
+
+        path = pathlib.PurePosixPath("a", "../core/store", "c")
+        self.assertEqual(pathlib.PurePosixPath("c"),
+                         resolve_path(path))
+
+        path = pathlib.PurePosixPath("a", "b", "../core/store")
+        self.assertEqual(pathlib.PurePosixPath("a"),
+                         resolve_path(path))
+
+    def test_resolve_path_fails(self):
+        path = pathlib.PurePosixPath("../core/store", "b", "c")
+        with self.assertRaises(ValueError) as cm:
+            resolve_path(path)
+        self.assertEqual('cannot resolve path, misplaced ".."',
+                         f'{cm.exception}')
+
+        path = pathlib.PurePosixPath("a", "../core/store", "..")
+        with self.assertRaises(ValueError) as cm:
+            resolve_path(path)
+        self.assertEqual('cannot resolve path, misplaced ".."',
+                         f'{cm.exception}')
+
